+++ cli/cli/kurtosis_gateway/server/engine_gateway/engine_gateway_service_server.go
@@ -113,15 +113,26 @@ func (service *EngineGatewayServiceServer) GetEnclaves(ctx context.Context, in *
 		return nil, stacktrace.Propagate(err, "An error occurred getting info for enclaves from the remote engine")
 	}
 	responseEnclaves := remoteEngineResponse.GetEnclaveInfo()
+	cleanUpRunningGateways := true
 	for enclaveId, enclaveInfo := range responseEnclaves {
 		var runningApiContainerGateway *runningApiContainerGateway
 		runningApiContainerGateway, isRunning := service.enclaveIdToRunningGatewayMap[enclaveId]
+		defer func() {
+			if cleanUpRunningGateways {
+				service.idempotentKillRunningGatewayForEnclaveId(enclaveId)
+			}
+		}()
 		// If the gateway isn't running, start it
 		if !isRunning {
 			runningApiContainerGateway, err = service.startRunningGatewayForEnclave(enclaveInfo)
+			if err != nil {
+				return nil, stacktrace.Propagate(err, "Expected to be able to start a local gateway for enclave '%v', instead a non-nil error was returned", enclaveId)
+			}
 		}
 		remoteEngineResponse.EnclaveInfo[enclaveId].ApiContainerHostMachineInfo = runningApiContainerGateway.hostMachineInfo
 	}
+
+	cleanUpRunningGateways = false
 	return remoteEngineResponse, nil
 }